Brainstorming file with some subheadings
This file will be available for anybody to add, at any time, ideas that
the have for the project. The included subheadings likely DO NOT cover
all aspects of the game. HOWEVER, if someone has an idea that does not
fall under any of these categores, they should still ADD THE IDEAS to
this file :)